requirements.txt
Flask
tensorflow
tf-keras
Pillow
numpy
gunicorn
